# openapi.yaml
openapi: 3.0.0
info:
  title: API de Charadas
  description: Uma API para gerenciar charadas (enigmas), incluindo autenticação administrativa e operações CRUD no Firestore.
  version: 1.0.0
  contact:
    name: João Pedro Stadler

servers:
  - url: https://charada-iota.vercel.app/
    description: Servidor Vercel (produção)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Insira o token JWT retornado no login. Exemplo: 'abc123...'"

  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é? Cai em pé e corre deitado."
        resposta:
          type: string
          example: "A chuva"
    
    Erro:
      type: object
      properties:
        error:
          type: string
          example: "Mensagem de erro explicativa"

paths:
  /:
    get:
      summary: Informações da API
      responses:
        '200':
          description: Dados básicos do autor e versão
          content:
            application/json:
              schema:
                type: object
                properties:
                  api: { type: string }
                  version: { type: string }
                  Author: { type: string }

  /login:
    post:
      summary: Autenticação de Administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usuario
                - senha
              properties:
                usuario:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  token: { type: string }
        '400':
          description: Dados ausentes
        '401':
          description: Usuário ou senha inválida

  /charadas:
    get:
      summary: Lista todas as charadas
      responses:
        '200':
          description: Uma lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'
    post:
      summary: Cria uma nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '201':
          description: Charada criada com sucesso
        '400':
          description: Dados inválidos ou falha no envio

  /charadas/aleatorias:
    get:
      summary: Retorna uma charada aleatória
      responses:
        '200':
          description: Uma única charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Busca charada pelo ID
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada
    
    put:
      summary: Atualização total da charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '200':
          description: Charada alterada com sucesso
        '404':
          description: Charada não encontrada

    patch:
      summary: Atualização parcial da charada
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '200':
          description: Charada alterada com sucesso
        '404':
          description: Charada não encontrada

    delete:
      summary: Remove uma charada
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Charada excluída com sucesso
        '404':
          description: Charada não encontrada